todolist/requirements.txt
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
cryptography==39.0.0 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
django-environ==0.9.0 ; python_version >= "3.10" and python_version < "4"
django-filter==22.1 ; python_version >= "3.10" and python_version < "4.0"
django==4.1.5 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
factory-boy==3.2.1 ; python_version >= "3.10" and python_version < "4.0"
faker==16.4.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
marshmallow-dataclass==8.5.11 ; python_version >= "3.10" and python_version < "4.0"
marshmallow==3.19.0 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.6.0 ; python_version >= "3.10" and python_version < "4.0"
pytest-django==4.5.2 ; python_version >= "3.10" and python_version < "4.0"
pytest-factoryboy==2.5.1 ; python_version >= "3.10" and python_version < "4.0"
pytest-faker==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
social-auth-app-django==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
social-auth-core==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspect==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
urllib3==1.26.13 ; python_version >= "3.10" and python_version < "4"
